migrate test utilities to go-libp2p-testing.
