--- ci/travis/travis_test.go
@@ -1,13 +1,0 @@
-package travis
-
-import (
-	"os"
-	"testing"
-)
-
-func TestIsRunning(t *testing.T) {
-	tr := os.Getenv("TRAVIS") == "true" && os.Getenv("CI") == "true"
-	if tr != IsRunning() {
-		t.Error("IsRunning() does not match TRAVIS && CI env var check")
-	}
-}
