+++ go.mod
@@ -1,8 +1,8 @@
 module github.com/libp2p/go-testutil
 
 require (
-	github.com/libp2p/go-libp2p-crypto v0.0.1
-	github.com/libp2p/go-libp2p-peer v0.0.1
-	github.com/multiformats/go-multiaddr v0.0.1
-	github.com/multiformats/go-multihash v0.0.1
+	github.com/libp2p/go-libp2p-core v0.0.0-20190508144953-ed42958fbb3a
+	github.com/libp2p/go-libp2p-testing v0.0.0-20190508145120-25bb910661fd
+	github.com/multiformats/go-multiaddr v0.0.2
+	golang.org/x/sys v0.0.0-20190222072716-a9d3bda3a223 // indirect
 )
